Create SinglyLinkedList.java
For this assignment, you will be implementing a Singly-Linked List with a head and tail reference. Recall that a Linked List is a collection of nodes, with each node containing a data item and reference(s) to other nodes (s). In a Singly-Linked List, each node will only reference the next node. Your Singly-Linked List should follow the requirements stated in the javadocs of each method you are to implement.

You will implement two add() methods. One will add to the front and one will add to the back. Be sure to correctly reassign the head and/or tail pointers when adding an element.

You will also implement two remove() methods. One will remove from the front and one will remove from the back. Be sure to correctly reassign the head and/or tail pointers when removing an element.